+++ anchor-image/src/main/java/org/anchoranalysis/image/voxel/extracter/ShortImplementation.java
@@ -76,7 +76,7 @@ class ShortImplementation extends Base<UnsignedShortBuffer> {
     }
 
     @Override
-    public int voxelWithMaxIntensity() {
+    public long voxelWithMaxIntensity() {
         return IterateVoxelsVoxelBoxAsInt.findMaxValue(voxels);
     }
